+++ example/schema.ts
@@ -1,14 +1,22 @@
-import { Elysia, t, SCHEMA } from '../src'
+import { Elysia, t, SCHEMA, DEFS } from '../src'
 
 const app = new Elysia()
-	// Strictly validate response
-	.get('/', () => 'hi', {
-		schema: {
+	.setModel(
+		'a',
+		t.Object({
 			response: t.String()
-		}
-	})
+		})
+	)
+	.setModel(
+		'b',
+		t.Object({
+			response: t.Number()
+		})
+	)
+	// Strictly validate response
+	.get('/', () => 'hi')
 	// Strictly validate body and response
-	.post('/', ({ body }) => body.id, {
+	.post('/', ({ body, query }) => body.id, {
 		schema: {
 			body: t.Object({
 				id: t.Number(),
@@ -16,10 +24,7 @@ const app = new Elysia()
 				profile: t.Object({
 					name: t.String()
 				})
-			}),
-			response: {
-				200: t.Number()
-			}
+			})
 		}
 	})
 	// Strictly validate query, params, and body
@@ -71,4 +76,7 @@ const app = new Elysia()
 	)
 	.listen(8080)
 
-type A = typeof app['store'][typeof SCHEMA]['/query/:id']['GET']['query']
+type A = typeof app['store'][typeof SCHEMA]['/']['POST']['query']
+type B = typeof app['store'][typeof DEFS]
+
+// const a = app.getModel('b')