+++ src/context.ts
@@ -1,16 +1,18 @@
 import { Elysia } from '.'
 import type { TypedRoute } from './types'
 
+type UnwrapFn<T> = T extends (...params: any) => any ? ReturnType<T> : T
+
 export interface Context<
 	Route extends TypedRoute = TypedRoute,
 	Store extends Elysia['store'] = Elysia['store']
 > {
 	request: Request
-	query: Route['query'] extends undefined
+	query: UnwrapFn<Route['query']> extends undefined
 		? Record<string, unknown>
-		: Route['query']
-	params: Route['params']
-	body: Route['body']
+		: UnwrapFn<Route['query']>
+	params: UnwrapFn<Route['params']>
+	body: UnwrapFn<Route['body']>
 	store: Store
 
 	set: {